requirements.txt
opencv-python==4.5.5.64
numpy==1.24.3
face-recognition==1.3.0
dlib>=19.7
pandas>=1.4.2
pillow>=9.0.0
